refactor: create locations/[uuid] endpoint, move opening hours to utils
